refactor: move functions to utils (#118)
- Expand filter column
- Extract flagged regions bed

So we can reuse them to extract the sample-specific flagged positions
bed.
e   (use "git push" to publish your local commits)
